Refactor visitor class to separate module, add unit tests
